+++ masci_tools/io/parsers/fleur_schema/schema_dict.py
@@ -35,7 +35,7 @@ from lxml import etree
 
 from masci_tools.util.lockable_containers import LockableDict, LockableList
 from masci_tools.util.case_insensitive_dict import CaseInsensitiveFrozenSet, CaseInsensitiveDict
-from masci_tools.util.xml.common_functions import abs_to_rel_xpath, split_off_tag
+from masci_tools.util.xml.common_functions import abs_to_rel_xpath, split_off_tag, contains_tag
 from .inpschema_todict import create_inpschema_dict, InputSchemaData
 from .outschema_todict import create_outschema_dict, merge_schema_dicts
 
@@ -364,7 +364,7 @@ class SchemaDict(LockableDict):
         paths = self._find_paths(name, self._tag_entries, contains=contains, not_contains=not_contains)
         #Filter out paths which contain the rootTag explicitly not a tag name containing the root tag
         #e.g. bravaisMatrix vs. bravaisMatrixFilm
-        paths = [path for path in paths if f'{root_tag}/' in path or path.endswith(root_tag)]
+        paths = [path for path in paths if contains_tag(path, root_tag)]
 
         relative_paths = {abs_to_rel_xpath(xpath, root_tag) for xpath in paths}
 
@@ -491,7 +491,7 @@ class SchemaDict(LockableDict):
         paths = self._find_paths(name, entries, contains=contains, not_contains=not_contains)
         #Filter out paths which contain the rootTag explicitly not a tag name containing the root tag
         #e.g. bravaisMatrix vs. bravaisMatrixFilm
-        paths = [path for path in paths if f'{root_tag}/' in path or path.endswith(root_tag)]
+        paths = [path for path in paths if contains_tag(path, root_tag)]
         relative_paths = {abs_to_rel_xpath(xpath, root_tag) for xpath in paths}
 
         if len(relative_paths) == 1:
@@ -945,15 +945,15 @@ class OutputSchemaDict(SchemaDict):
             raise ValueError(f"{iteration_tag} is not a valid iteration tag valid are: {list(self['iteration_tags'])}")
         iteration_path = self.tag_xpath(iteration_tag)
 
-        if f'/{root_tag}' not in iteration_path:
+        if not contains_tag(iteration_path, root_tag):
             #The paths have to include the root_tag
             contains = _add_condition(contains, root_tag)
 
         paths = self._find_paths(name, ('iteration_tag_paths',), contains=contains, not_contains=not_contains)
-        if f'/{root_tag}' not in iteration_path:
+        if not contains_tag(iteration_path, root_tag):
             #Filter out paths which contain the rootTag explicitly not a tag name containing the root tag
             #e.g. bravaisMatrix vs. bravaisMatrixFilm
-            paths = [path for path in paths if f'{root_tag}/' in path or path.endswith(root_tag)]
+            paths = [path for path in paths if contains_tag(path, root_tag)]
         paths = [f"{iteration_path}{path.lstrip('.')}" for path in paths]
         relative_paths = {abs_to_rel_xpath(xpath, root_tag) for xpath in paths}
 
@@ -1102,14 +1102,14 @@ class OutputSchemaDict(SchemaDict):
             if 'iteration' in entry and all(f'{excl}_attribs' not in entry for excl in exclude)
         ]
 
-        if f'/{root_tag}' not in iteration_path:
+        if not contains_tag(iteration_path, root_tag):
             contains = _add_condition(contains, root_tag)
 
         paths = self._find_paths(name, entries, contains=contains, not_contains=not_contains)
-        if f'/{root_tag}' not in iteration_path:
+        if not contains_tag(iteration_path, root_tag):
             #Filter out paths which contain the rootTag explicitly not a tag name containing the root tag
             #e.g. bravaisMatrix vs. bravaisMatrixFilm
-            paths = [path for path in paths if f'{root_tag}/' in path or path.endswith(root_tag)]
+            paths = [path for path in paths if contains_tag(path, root_tag)]
         paths = [f"{iteration_path}{path.lstrip('.')}" for path in paths]
         relative_paths = {abs_to_rel_xpath(xpath, root_tag) for xpath in paths}